Refactor base app completely
Remove view for other sims till they are upgraded.
