+++ index.html
@@ -38,6 +38,8 @@
       </p>
       <span class="card-tag">Custom Sim</span>
     </a>
+
+    <!--
     <a class="sim-card" href="profit-loss-sim.html">
       <h2>Profit &amp; Loss Simulation</h2>
       <p>
@@ -79,7 +81,7 @@
       </p>
       <span class="card-tag">International Retirement Sim</span>
     </a>
-
+   -->
   </div>
 
 </body>
